+++ src/app/user/[id]/setting/edit/components/EditSex/EditSex.tsx
@@ -1,35 +1,36 @@
 import { UseFormRegister, UseFormRegisterReturn } from "react-hook-form";
 import * as GS from "../../UserEdit.styles";
-import * as S from "./EditAge.styles";
+import * as S from "./EditSex.styles";
 import { UserEditData } from "../../UserEdit.types";
 interface EditAgeProps {
   register: UseFormRegister<UserEditData>;
 }
-const EditAge = ({ register }: EditAgeProps) => {
+const EditSex = ({ register }: EditAgeProps) => {
   const options = [
     { text: "남성", value: "male" },
     { text: "여성", value: "female" },
   ];
 
   return (
-    <S.EditAgeContainer>
+    <S.EditSexContainer>
       <GS.UserEditTitle>성별</GS.UserEditTitle>
 
-      <S.EditAgeActions>
+      <S.EditSexActions>
         {options.map(({ text, value }) => (
           <InputRadio
             key={`${text}_${value}`}
             value={value}
             text={text}
+            isSelected={true}
             register={register("sex", { required: "성별의 입력은 필수" })}
           />
         ))}
-      </S.EditAgeActions>
-    </S.EditAgeContainer>
+      </S.EditSexActions>
+    </S.EditSexContainer>
   );
 };
 
-export default EditAge;
+export default EditSex;
 
 interface createInputRadioProps {
   register: UseFormRegisterReturn;
