+++ src/app/user/[id]/setting/edit/components/index.ts
@@ -1,2 +1,2 @@
 export { default as EditNickname } from "./EditNickname/EditNickname";
-export { default as EditAge } from "./EditAge/EditAge";
+export { default as EditSex } from "./EditSex/EditSex";
